+++ src/main/java/com/example/vpapi/service/ImageServiceImpl.java
@@ -3,6 +3,7 @@ package com.example.vpapi.service;
 import com.example.vpapi.domain.Image;
 import com.example.vpapi.dto.ImageDTO;
 import com.example.vpapi.repository.ImageRepository;
+import com.example.vpapi.util.CustomServiceException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.stereotype.Service;
@@ -19,7 +20,7 @@ public class ImageServiceImpl implements ImageService {
     @Override
     public ImageDTO get(Long ino) {
         Optional<Image> result = imageRepository.findById(ino);
-        Image image = result.orElseThrow();
+        Image image = result.orElseThrow(() -> new CustomServiceException("NOT_EXIST_IMAGE"));
         return entityToDTO(image);
     }
 
@@ -32,6 +33,9 @@ public class ImageServiceImpl implements ImageService {
 
     @Override
     public void remove(Long ino) {
+        if (!imageRepository.existsById(ino)) {
+            throw new CustomServiceException("NOT_EXIST_IMAGE");
+        }
         imageRepository.deleteById(ino);
     }
