+++ src/main/java/com/example/vpapi/service/VideoServiceImpl.java
@@ -3,6 +3,7 @@ package com.example.vpapi.service;
 import com.example.vpapi.domain.Video;
 import com.example.vpapi.dto.VideoDTO;
 import com.example.vpapi.repository.VideoRepository;
+import com.example.vpapi.util.CustomServiceException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.stereotype.Service;
@@ -19,7 +20,7 @@ public class VideoServiceImpl implements VideoService {
     @Override
     public VideoDTO get(Long vno) {
         Optional<Video> result = videoRepository.findById(vno);
-        Video video = result.orElseThrow();
+        Video video = result.orElseThrow(() -> new CustomServiceException("NOT_EXIST_VIDEO"));
         return entityToDTO(video);
     }
 
@@ -32,6 +33,9 @@ public class VideoServiceImpl implements VideoService {
 
     @Override
     public void remove(Long vno) {
+        if (!videoRepository.existsById(vno)) {
+            throw new CustomServiceException("NOT_EXIST_VIDEO");
+        }
         videoRepository.deleteById(vno);
     }
